feat: reduce amount of rpc calls
