+++ src/health/health.ts
@@ -1,7 +1,3 @@
-import express from 'express';
-import asyncHandler from 'express-async-handler';
-import { MAIN_CONFIG } from '../config';
-import { createLogger } from '../util/logger';
 import { getAllInstalledSolutionsNames, runtimeStarted } from '../node-red/red';
 
 enum HealthStatus {
@@ -13,37 +9,6 @@ enum ComponentName {
   RED = 'NODE_RED',
   READY = 'READY',
 }
-
-export const createHealthRouter = (): express.Router | null => {
-  if (!MAIN_CONFIG.ENABLE_HEALTH_API) {
-    return null;
-  }
-
-  const healthLogger = createLogger('Health');
-
-  const healthRouter: express.Router = express.Router({ mergeParams: true });
-
-  healthRouter.get('/health/liveness', (req, res) => {
-    const health = isLive();
-
-    healthLogger.debug('requested liveness');
-
-    res.json(health);
-  });
-
-  healthRouter.get(
-    '/health/readiness',
-    asyncHandler(async (req, res) => {
-      const result = await isReady();
-
-      healthLogger.debug(result, 'requested readiness');
-
-      res.json(result);
-    }),
-  );
-
-  return healthRouter;
-};
 
 interface ComponentHealthStatus {
   status: HealthStatus;
@@ -58,14 +23,14 @@ interface NodeRedHealthStatus extends ComponentHealthStatus {
   };
 }
 
-const isLive = (): ComponentHealthStatus => {
+export const isLive = (): ComponentHealthStatus => {
   return {
     status: HealthStatus.OK,
     name: 'LIVE',
   };
 };
 
-const isReady = async (): Promise<ComponentHealthStatus[]> => {
+export const isReady = async (): Promise<ComponentHealthStatus[]> => {
   return [await getNodeRedHealth()];
 };
