+++ src/solution.ts
@@ -33,12 +33,13 @@ const logger = createLogger('SolutionLoop');
 
 export const pushToQueue = async (account: KeyringPair): Promise<void> => {
   // eslint-disable-next-line no-constant-condition
-  const api: ApiPromise = await retryHttpAsyncCall(async () => await createReadPalletApi());
 
   const timeout: number = MAIN_CONFIG.SOLUTION_QUEUE_PROCESS_DELAY;
 
   // eslint-disable-next-line no-constant-condition
   while (true) {
+    const api: ApiPromise = await retryHttpAsyncCall(async () => await createReadPalletApi());
+
     await processSolutionQueue(api, account).catch(async (e) => {
       logger.error('failed to complete queue loop');
       logger.error(e);
@@ -102,7 +103,7 @@ async function processSolutionQueue(api: ApiPromise, workerAccount: KeyringPair)
     operatorSubscriptions,
   );
 
-  const unfilteredSolutions: SolutionArray = await getSolutions(api);
+  const unfilteredSolutions: SolutionArray = await getSolutions(api, operatorSubscriptions);
 
   const solutions: SolutionArray = unfilteredSolutions.filter((solution) =>
     operatorSubscriptions.includes(solution[1]),
@@ -146,16 +147,34 @@ async function processSolutionQueue(api: ApiPromise, workerAccount: KeyringPair)
     return;
   }
 
+  const solutionGroupStatus = new Set<string>();
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  for (const [_, solutionGroup] of Object.entries(solutionGroups)) {
+    const hasValidConfiguration: boolean = await hasValidGroupConfiguration(
+      api,
+      operatorAddress,
+      solutionGroup,
+    );
+
+    if (!hasValidConfiguration) {
+      logger.warn(
+        { solutionGroupId: solutionGroup.namespace },
+        'operator does not meet criteria for solution group, skipping installation',
+      );
+
+      continue;
+    }
+
+    solutionGroupStatus.add(solutionGroup.namespace);
+  }
+
   for (const solution of activeTargetSolutions) {
     const workLogic: string = solution[2].workload.workLogic;
 
-    const isSuccesful: boolean = await hasValidGroupConfiguration(
-      api,
-      operatorAddress,
-      solutionGroups[solution[1]],
-    );
+    const fulfillsSolutionGroupCriteria = solutionGroupStatus.has(solution[1]);
 
-    if (!isSuccesful) {
+    if (!fulfillsSolutionGroupCriteria) {
       logger.warn(
         {
           solutionId: solution[0],
